+++ src/platform/src/Bridge/Failover/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.7
+---
+
+ * Add support for model overrides per platform
+ * Add support for multiple model fallbacks per platform entry
+
 0.2
 ---